requirements.txt
# Victor Bot dependencies

discord.py>=2.3.2
